requirements-runpod.txt
# VantageVeoConverter RunPod Serverless Requirements
# Optimized for RunPod deployment with minimal dependencies

# RunPod SDK (REQUIRED)
runpod>=1.0.0

# Core numerical libraries
numpy>=1.24,<2.3
scipy>=1.9.0

# PyTorch (pre-installed in RunPod base image, but specified for completeness)
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Computer Vision
opencv-python>=4.5.0
scikit-image>=0.19.0

# Audio Processing
openai-whisper>=20230314
librosa>=0.9.0

# Audio alignment (install with --no-build-isolation)
# aeneas>=1.7.3

# Essential utilities
tqdm>=4.64.0
Pillow>=9.0.0
requests>=2.25.0
numba>=0.56.0

# Optional performance boosters
psutil>=5.8.0
ffmpeg-python>=0.2.0

# For file handling in serverless
boto3>=1.26.0